using custom iter_futures to avoid dups code
